Refactor: Move qualifiers back to core:common and fix modularization layering
- Created CommonQualifiers.kt in :core:common with KDoc for all qualifier annotations.
- Added javax.inject to :core:common and removed :data:common dependency from :core:camera.
- Kept only Hilt @Module in :data:common/CommonModule.kt.
- Updated import statements across consumer modules to com.google.jetpackcamera.core.common.*.
