+++ data/common/consumer-rules.pro
@@ -1,0 +1,1 @@
+# Proguard rules for data:common
